requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.2.1

# AI/ML dependencies
openai==1.10.0
sentence-transformers==5.1.2
faiss-cpu==1.9.0.post1
numpy==2.3.4

# Data validation
pydantic==2.5.3

# Testing
pytest==8.0.0
httpx==0.26.0
pytest-asyncio==0.23.3

# Development
black==24.1.1
flake8==7.0.0
